restructure loader
